Remove unneeded extra element
